+++ domain/src/main/java/tachiyomi/domain/library/service/LibraryPreferences.kt
@@ -54,8 +54,6 @@ class LibraryPreferences(
         ),
     )
 
-    val autoUpdateMetadata: Preference<Boolean> = preferenceStore.getBoolean("auto_update_metadata", false)
-
     val showContinueReadingButton: Preference<Boolean> = preferenceStore.getBoolean(
         "display_continue_reading_button",
         false,